requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.2
django-admin-confirm==0.2.5
django-admin-list-filter-dropdown==1.0.3
django-auto-logout==0.5.0
django-filter==21.1
django-import-export==2.7.1
django-session-timeout==0.1.0
django-simpleui==2022.6.131
django-environ
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
lxml==4.8.0
MarkupPy==1.14
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
Pillow==9.1.0
portalocker==2.4.0
protobuf==4.21.2
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyMySQL==1.0.2
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
#pywin32==304
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.1.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
